chore: remove example classes and reorganize package
