--- server/src/site/doget/dto/BankDto.java
@@ -1,12 +1,0 @@
-package site.doget.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class BankDto {
-
-    int code;
-    String name;
-}
