--- server/src/site/doget/mybatis/mapper/BankMapper.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-
-<mapper namespace="site.doget.repository.BankRepository">
-    
-    <resultMap id="BankDtoResultMap" type="bankDto">
-        <result property="code" column="code"/>
-        <result property="name" column="name"/>
-    </resultMap>
-    <select id="findByCode" resultType="bankDto" parameterType="int">
-        <![CDATA[
-            SELECT * FROM bank_table WHERE code=#{code}
-        ]]>
-    </select>
-
-</mapper>
